Update to run a specified number of customers in separate processes that controls a bike object.
